+++ dtoolcore/__init__.py
@@ -302,6 +302,13 @@ class ProtoDataSet(_BaseDataSet):
         for handle in self._storage_broker.iter_item_handles():
             yield dtoolcore.utils.generate_identifier(handle)
 
+    def update_name(self, new_name):
+        """Update the name of the proto dataset.
+
+        :param new_name: the new name of the proto dataset
+        """
+        self._admin_metadata['name'] = new_name
+
     def create(self):
         """Create the required directory structure and admin metadata."""
         self._storage_broker.create_structure()